backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
filelock==3.19.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.7.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.15.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lxml==6.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numpy==1.24.3
ollama==0.5.3
packaging==25.0
pandas==2.0.3
pdfminer.six==20221105
pdfplumber==0.9.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.3.0
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
transformers==4.33.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Wand==0.6.13
Werkzeug==3.1.3
